+++ src/OmniSharp.Abstractions/Models/v1/FixAll/RunFixAllScope.cs
@@ -1,0 +1,9 @@
+namespace OmniSharp.Abstractions.Models.V1.FixAll
+{
+    public enum FixAllScope
+    {
+        Document = 0,
+        Project = 1,
+        Solution = 2
+    }
+}